requirements.txt
streamlit
faiss-cpu
sentence-transformers
numpy
python-dotenv
groq
tavily-python
pypdf